Remove requirements file for building docs
The preferred method of building the docs is through tox:

    $ python -m tox -e docs

There is no need for installing Sphinx or any other library directly.

On a semi-related note, the Read the Docs theme is no longer a
dependency of Sphinx and now needs to be included as a separate library
to install.