Disable Sentry when no DSN is provided
If an application tried to use Henson-Sentry without specifying a DSN,
`Sentry` would raise an exception (by way of `Extension`). This meant
that either a value was required in all environments (e.g., production,
staging) or code needed to be changed between environments.

A better approach is for Henson-Sentry to gracefully disable itself when
no DSN is provided. When that happens, a message will be logged in case
the behavior wasn't expected.